server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.1.0
httpx==0.24.1
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic-settings==2.1.0
python-magic==0.4.27
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
openai>=1.30.0
numpy<2.0.0
torch==2.1.0
torchaudio==2.1.0
pyannote-audio==3.1.1 